quit in trap only for non-zero exit code

+++ contrib/scripts/load-test.sh
@@ -5,8 +5,10 @@ set -e
 source contrib/scripts/functions.sh
 
 function finish {
-	quit 0
-	rm -rf $1
+  if [ $? -ne 0 ]; then
+	  quit 0
+	  rm -rf $1
+  fi
 }
 
 trap finish EXIT
